Don't singularize the type
Depending on your structure this might not be the right thing to do.
If you name your ability `posts.js` or using the pod structyre it
lives in `posts/ability.js` the helper will be `{{if (can "write posts")}}`

If you name your file in a singular way it will be `{{if (can "write post"}}`
as it used to be.